fix bug

+++ src/model/trainer.py
@@ -32,7 +32,7 @@ class Trainer:
             for src,target_input, target_output, target_padding, output_padding in self.dataloader:
                 logits         = self.model(src,target_input,target_padding) # (B,L,V)
                 logits         = logits.reshape(logits.shape[0] * logits.shape[1],logits.shape[2])
-                output_padding = output_padding.reshape(-1)
+                output_padding = output_padding.reshape(-1).int()
                 loss           = self.criterion(logits[output_padding!=0],target_output[output_padding!=0])
                 self.optimizer.zero_grad()
                 loss.backward()
